力扣：437_路径总和III

+++ src/main/java/com/leetcode/LeetCode_437_路径总和III_12_20.java
@@ -9,13 +9,13 @@ public class LeetCode_437_路径总和III_12_20 {
 
     public int pathSum(TreeNode root, int targetSum) {
         // 和 560 题一样套路，使用前缀和
-        Map<Integer, Integer> map = new HashMap<>();
-        map.put(0, 1);
-        dfs(root, 0, targetSum, map);
+        Map<Long, Integer> map = new HashMap<>();
+        map.put(0L, 1);
+        dfs(root, 0L, targetSum, map);
         return ans;
     }
 
-    private void dfs(TreeNode node, int s, int targetSum, Map<Integer, Integer> map) {
+    private void dfs(TreeNode node, Long s, int targetSum, Map<Long, Integer> map) {
         if (node == null) {
             return;
         }
